toSeconds fix

+++ madman/src/main/java/com/flipkart/madman/helper/Utils.kt
@@ -95,7 +95,7 @@ object Utils {
         }
     }
 
-    fun formatSecondsToMMSS(seconds: Int): String {
+    fun formatSecondsToMMSS(seconds: Long): String {
         val minutes = (seconds % 3600) / 60
         val remainingSeconds = seconds % 60
         return "%02d:%02d".format(minutes, remainingSeconds)
